+++ spec/unit/generator/embed_frameworks_script_spec.rb
@@ -4,7 +4,7 @@ module Pod
   describe Generator::EmbedFrameworksScript do
     it 'returns the embed frameworks script' do
       frameworks = {
-        'Debug' => [Target::FrameworkPaths.new('Pods/Loopback.framework', 'Pods/Loopback.framework.dSYM'),
+        'Debug' => [Target::FrameworkPaths.new('Pods/Loopback.framework', 'Pods/Loopback.framework.dSYM', ['7724D6B4-C7DD-31F0-80C6-EE818ED30B07.bcsymbolmap', 'B724D6B4-C7DD-31F0-80C6-EE818ED30B0B.bcsymbolmap']),
                     Target::FrameworkPaths.new('Reveal.framework')],
         'Release' => [Target::FrameworkPaths.new('CrashlyticsFramework.framework')],
       }
@@ -13,6 +13,8 @@ module Pod
         if [[ "$CONFIGURATION" == "Debug" ]]; then
           install_framework "Pods/Loopback.framework"
           install_dsym "Pods/Loopback.framework.dSYM"
+          install_bcsymbolmap "7724D6B4-C7DD-31F0-80C6-EE818ED30B07.bcsymbolmap"
+          install_bcsymbolmap "B724D6B4-C7DD-31F0-80C6-EE818ED30B0B.bcsymbolmap"
           install_framework "Reveal.framework"
         fi
       SH