+++ CHANGELOG.md
@@ -8,6 +8,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Enhancements
 
+* Copies the bcsymbolmap files of a vendored framework  
+  [dacaiguoguogmail](https://github.com/dacaiguoguogmail)
+  [#8470](https://github.com/CocoaPods/CocoaPods/pull/8470)
+
 * Set the path of development pod groups to root directory of the Pod 
   [Eric Amorde](https://github.com/amorde)
   [#8445](https://github.com/CocoaPods/CocoaPods/pull/8445)